fix: методы отправки и приема запросов/ответов объединены в один класс
