+++ modules/client/src/test/java/io/deeplay/intership/client/AuthorizationControllerTest.java
@@ -1,6 +1,6 @@
 package io.deeplay.intership.client;
 
-import io.deeplay.intership.connection.ClientStreamConnector;
+import io.deeplay.intership.connection.StreamConnector;
 import io.deeplay.intership.ui.UserInterface;
 import io.deeplay.intership.decision.maker.DecisionMaker;
 import io.deeplay.intership.decision.maker.LoginPassword;
@@ -19,7 +19,7 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 public class AuthorizationControllerTest {
-    private final ClientStreamConnector streamConnector = mock(ClientStreamConnector.class);
+    private final StreamConnector streamConnector = mock(StreamConnector.class);
     private final UserInterface userInterface = mock(UserInterface.class);
     private final DecisionMaker decisionMaker = mock(DecisionMaker.class);
